Bug fix---uuidlib_toa() needs to return an int, not a size_t.

+++ src/uuid.h
@@ -80,6 +80,6 @@ int	uuidlib_v4		(uuid__t *const)                                                
 int	uuidlib_v5		(uuid__t *const,const uuid__t *const,const void *const,const size_t) __attribute__((nonnull,nothrow));
 int	uuidlib_cmp		(const uuid__t *const restrict,const uuid__t *const restrict)        __attribute__((nonnull,nothrow));
 int	uuidlib_parse		(uuid__t *const,const char *)                                        __attribute__((nonnull,nothrow));
-size_t	uuidlib_toa		(const uuid__t *const,char *dest,size_t)                             __attribute__((nonnull,nothrow));
+int	uuidlib_toa		(const uuid__t *const,char *dest,size_t)                             __attribute__((nonnull,nothrow));
 
 #endif
